backend/requirements_new.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
pandas==2.1.1
joblib==1.3.2
scikit-learn==1.3.0
PyMySQL==1.1.0
cryptography==41.0.4
mysql-connector-python==8.1.0